create AbstractGit
