+++ tests/Unit/GitFactoryMethodsTest.php
@@ -88,6 +88,8 @@ class GitFactoryMethodsTest extends GitTestCase
      * @covers \ArtARTs36\GitHandler\Git::attributes
      * @covers \ArtARTs36\GitHandler\Git::submodules
      * @covers \ArtARTs36\GitHandler\Git::backup
+     * @covers \ArtARTs36\GitHandler\Git::createBackupElements
+     * @covers \ArtARTs36\GitHandler\Git::createSubjectsConfigurators
      */
     public function testFactoryMethod(string $method, string $expectedClass): void
     {
